ai-recipes/07-multi-agent/openai-sdk/requirements.txt
durabletask>=0.3.0
durabletask-azuremanaged>=0.3.0
openai>=1.60.0
azure-identity>=1.17.0
pydantic>=2.0
python-dotenv>=1.0.0
